Maintenance: Rename TokenInfoPopover folder to TokenInfo
